tests/wpt/tests/css/css-values/tree-counting/sibling-index-keyframe-font-style-dynamic.html

<!DOCTYPE html>
<title>CSS Values and Units Test: sibling-index() changing font-style during @keyframes animation</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#tree-counting">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes --anim {
    from {
      font-style: oblique calc(5deg * sibling-index());
    }
    to {
      font-style: oblique 9deg;
    }
  }
  #target {
    animation: --anim 1000s step-end;
  }
</style>
<div>
  <div id="rm"></div>
  <div></div>
  <div id="target"></div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(target).fontStyle, "oblique 15deg");
  }, "Initially, the sibling-index() is 3 for #target");

  test(() => {
    rm.remove();
    assert_equals(getComputedStyle(target).fontStyle, "oblique 10deg");
  }, "Removing a preceding sibling of #target reduces the sibling-index()");
</script>
